+++ packages/web-components/src/link/define-async.ts
@@ -1,5 +1,4 @@
-import { RenderableFASTElement } from '@microsoft/fast-html';
-import { definition } from './link.definition-async.js';
+import { declarativeDefinition } from './link.definition-async.js';
 import { Link } from './link.js';
 
-RenderableFASTElement(Link).defineAsync(definition);
+Link.define(declarativeDefinition);